Fixed Bench module is redundant #6
